LayoutTests/canvas/philip/tests/2d.drawImage.zerosource-expected.txt

< [index] >
2d.drawImage.zerosource
drawImage with zero-sized source rectangle throws INDEX_SIZE_ERR
References: 2d.drawImage.zerosource
Actual output:
Expected output:

Passed
